user_managment/srcs/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.0.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-redis==5.4.0
gunicorn==20.1.0
PyJWT==2.10.1
python-decouple==3.8
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Pillow==10.0.0
whitenoise==6.6.0
watchdog